fix: provide common component Navigation
